file path

+++ zikeiretsu/src/tsdb/query/lexer/mod.rs
@@ -13,7 +13,7 @@ use either::Either;
 use serde::{Deserialize, Serialize};
 use std::fs;
 use std::io::Error as IoError;
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::result::Result as StdResult;
 use thiserror::Error;
 
@@ -50,6 +50,9 @@ pub type Result<T> = std::result::Result<T, LexerError>;
 pub enum OutputError {
     #[error("{0}")]
     IoError(#[from] IoError),
+
+    #[error("output file already exists:{0}")]
+    OutputFileAlreadyExists(PathBuf),
 
     #[error("invalid output file path : {0}")]
     InvalidPath(String),
@@ -111,7 +114,7 @@ impl OutputWriter {
         match format {
             OutputFormat::Json => Ok(()),
             OutputFormat::Table => Ok(()),
-            OutputFormat::Parquet => match &self {
+            OutputFormat::Parquet | OutputFormat::ParquetSnappy => match &self {
                 OutputWriter::File(_) => Ok(()),
                 OutputWriter::Stdout => Err(OutputError::InvalidOutputDestination(
                     "parquet format can output to only a file".to_string(),
@@ -129,11 +132,18 @@ impl OutputCondition {
             Some(output_file_path) => match output_file_path.parent() {
                 None => Err(OutputError::InvalidPath(format!("{:?} ", output_file_path))),
                 Some(output_dir) => {
-                    if output_dir.exists() {
-                        fs::create_dir_all(output_dir)?;
+                    if Path::new(output_file_path).exists() {
+                        Err(OutputError::OutputFileAlreadyExists(
+                            output_file_path.clone(),
+                        ))
+                    } else {
+                        if output_dir.exists() {
+                            fs::create_dir_all(output_dir)?;
+                        }
+
+                        let f = fs::File::create(output_file_path)?;
+                        Ok(OutputWriter::File(f))
                     }
-                    let f = fs::File::create(output_file_path)?;
-                    Ok(OutputWriter::File(f))
                 }
             },
         }?;
